backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
torch==1.13.1
torchvision==0.14.1
pillow==11.1.0
matplotlib==3.8.0
gunicorn
numpy==1.26.4
